Changed the number of decimal places printed from fixed to variable.
I changed the number of decimal places printed for a failure from a
fixed "3" to the precision being tested +1. 

This will allow it to work with any other level of precision that is
passed in to the function.  In addition it will show errors where they
are only different by a rounded part (causing the assertion, but looking
the same).  For example 3.1415 and 3.1414 will throw an assertion (round
to 3.142 and 3.141) but will both print as 3.141 with %.3f.  ...%." +
str(precision) + "f... will print the extra digit that determines
rounding.